micro: Printed memory plan layout changes
Reformats output of GreedyMemoryPlanner:PrintMemoryPlan() to make it
simpler to interpret.

Per-buffer info line improvements:

(a) Reduces text quantity to make it easier to scan for information.
(b) Moves size to near front line, to make it simpler for the eye to
    find this information.
(c) Includes ordinal letter used in per-time display below to make it
    simpler to cross-reference.

Per time improvements:

(a) Includes tick number. This is useful in larger models for cross
    referencing to the per-buffer information and also for helping to
    determine the actual operation being executed at that time.
(b) Includes total memory use of buffers, which helps to more clearly
    identify memory bottlenecks.