[JAX] Include both the global and thread-local extra jit context in the JIT signature.
Previously if there was a thread-local context set we would include it in preference to the global context. However since the "extra context" is an opaque object with no particular semantics as far as the C++ code is concerned, it makes more sense to include both local and global as part of the cache key. Just because the thread-local values are equal it doesn't mean that they completely override the global values.

PiperOrigin-RevId: 369303867
Change-Id: Id7172fb71f925f62f6fa3f8f596e12b635734dd0