Variable restore-on-create from CheckpointInitialValue: don't call `set_shape` on the restored tensor.
restore-on-create always runs eagerly under init_scope, therefore I believe the restored tensor should always have a shape. So the purpose of set_shape is unclear. Also `EagerTensor.set_shape` [1] doesn't really set the shape: it just checks if the given shape is compatible with tensor's shape. This check happens elsewhere as well (e.g, variables creation [2], variable assign ops) so we are not losing this check.

The problem with `set_shape` is that it fetches the tensor's shape, which blocks until the tensor is ready. For async eager, this is a synchronization point and makes it impossible to parallelize multiple checkpoint restoration w/o resorting to Python threads. This slows down the restoration of sharded variable in TPUMPStrategy (we restore each shard inside a python loop [3]). This slowdown is quite significant if number of shards is large (e.g, 128).

PiperOrigin-RevId: 369948633
Change-Id: Id538ea8ee3f8fc1b50e324b8c4308f27ed9e4957