[XLA] Fix cost analysis interval picking when there is no prefetch start time that satisfies earliest < x < latest.
We specify how long a prefetch can be in relation to how long the overlapped
instructions take using flags. However, sometimes there is a long-executing HLO
before the HLO and the earliest and latest durations both fall within this HLO.
In this case, we were previously not attempting to prefetch at all because there
is no valid prefetch start time that satisifies the earliest/latest constraints.
However, this can be detrimental to some models. We now allow prefetches to
start a little earlier than the specified earliest time in such cases.

PiperOrigin-RevId: 370087066
Change-Id: Ief3b6cd58cdf6d26b38569a5a541c8e06b946840