[tf.data] Improve error message for iterator restoration failure.
Previously, trying to restore into a modified dataset would raise an error such as "NotFoundError: Iterator::Model::TensorSlice". This would get caught in tracking/util.py and be incorrectly reraised as an error saying that the checkpoint couldn't be found.

With this CL we provide a more accurate and actionable message:

Failed to restore from checkpoint or SavedModel at <path>: Failed to restore dataset iterator from checkpoint: Iterator::Model::TensorSlice. Make sure the dataset definition has not changed between the process that saved the checkpoint and the process that is restoring it. [Op:DeserializeIterator]

PiperOrigin-RevId: 370559855
Change-Id: I56c55fd7cf098d52da03994179c5acc36a14fb3a