[tf.data] Refactoring of dataset finalization.
This CL introduces a new utility method `FinalizeDataset`, which is intended to be applied to `DatasetBase` representing the final transformation of a user-defined input pipeline before creating an iterator for it. The method ends up performing static graph optimizations and appends a new `Root` transformation that combines the functionality of existing internal transformations -- `ModelDataset`, `MaxIntraOpParallelismDataset`, and `PrivateThreadpoolDataset`.

This CL also performs refactoring of `FinalizeDatasetOp` and `OptimizeDatasetOp` to avoid duplication of logic and switches the `tf.data.experimental.save()` and `tf.data.experimental.TFRecordWriter()` APIs to use the new finalization mechanism, replacing the previous mechanism based on `_apply_options()` and `FinalizeDatasetOp`.

Subsequent changes will make a similar switch for other mechanisms that iterate through a tf.data dataset, such as `tf.data.Iterator` or `tf.data.Dataset.reduce()`.

PiperOrigin-RevId: 371712364
Change-Id: Ide29cc2713a0a4f9194f1ba4bc5840aa73dbe5c4