Re-introduce eliminate pass through arguments pattern by limiting it to strings
This will closely match the old bridge behavior which moves out such return values while moving identity nodes to host during outside compilation. So, this is less likely to introduce unintended consequences in other parts of the stack.

We can fully generalize this pattern once the XLA issue is fixed.

PiperOrigin-RevId: 372236912
Change-Id: Ib4beb64d5183ccc22e900780ef965e1809a30add