Fix incorrect dependency on mlir::createLowerAffinePass
Years ago this pass used to be defined in mlir/Transforms, but is now located in Conversion/AffineToStandard. Unfortunately the forward declaration of the create method was left in mlir/Transforms/Passes.h, which allowed for these incorrect dependencies to still "work". This revision corrects the dependency now that the stale forward declaration of createLowerAffinePass is being removed.

PiperOrigin-RevId: 374234918
Change-Id: Ie4e18c4ec8a8e70c9e667d20d52143f349e1d212