[tf:jit] Resolve statically known constraints during clustering
Stop constraints propagation early if they can be resolved statically from tensor type. This is mostly a performance improvement change to stop large IR traversals early, if we know for sure that all constraints will be satisfied.

As a side effect only unresolved constraints will be added to the compiled function signature, which makes it easier to read.

PiperOrigin-RevId: 374508333
Change-Id: Id62c8b91407243181c992a9d7f1df33df0dbe7a6