[SE:GPU] Change code that invokes ptxas to prefer ptxas from PATH rather than from hard-coded locations like /usr/local/cuda.
We determine if a valid ptxas is on PATH by trying to run `ptxas --version`. If that works, use it.

The advantage of looking in PATH first is that users can override the PATH, whereas they may not have control over, say, an old version of CUDA in /usr/local/cuda on a shared machine.

Add caching around code that finds executables. There's no need to access the filesystem repeatedly.

Also lower verbosity of failed Subprocess::Start() calls. With this change, we expect some subprocesses starts to fail, so we don't want to LOG about them.

https://github.com/google/jax/issues/6799

PiperOrigin-RevId: 375104404
Change-Id: I269a672cc3fdcfdc3c39ec1be025b6f326f7ae89