Get rid of explicit D2H copies in GPU NMS implementation.
Instead, have the GPU write directly to pinned CPU memory.

The GPU can do only one D2H memcopy at once.  If there is already a D2H memcopy
running, we have to wait until it's complete before we can do the one here.

But if we write directly to pinned memory, we don't have this problem.

We've observed this technique elsewhere lead to large latency reductions in
contended cases.

PiperOrigin-RevId: 375181403
Change-Id: I0602cb20e8b0d7536a41cb361f6401a436b93285