Extend scalar broadcast optimization for ConstOp to all types
ConstOp xla op kernel avoids creation of large tensors if the proto representation has only one element. This encodes tensors that have the same value repeated for all the elements. This was limited to only some of the types supported by XLA.

This change extends the above optimization to all the primitive types supported by XLA.

Added a distribution strategy test as we want to feed the full program to XLA in the test to make sure the large tensor is not materialized anywhere.

PiperOrigin-RevId: 375543785
Change-Id: I7ca59bef59f39a3f8e5a5c1e8c83ba11bd227025